update_url for generating installable .crx
To generate a installable .crx file, which will not show as corrupted in the browser, update_url is needed. However, I'm not using chrome developer dashboard or other host servers like that, so I'd like to add a dummy url at this point.
If the extension is going to the store, remember to change update_url otherwise the updating procedure will be interfered.